Time: 2 ms (95.99%), Space: 46.1 MB (8.52%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.java
@@ -19,7 +19,7 @@ class Solution {
         queue.add(root);
         while(!queue.isEmpty()){
             int size = queue.size();
-            Double sum = (double)0;
+            double sum = 0;
             for(int i = 0; i < size; i++){
                 TreeNode curr = queue.poll();
                 sum += curr.val;
